Containerise the thank you page
