+++ assets/pages/oneoff-contributions/components/contributionsThankYouPageContainer.jsx
@@ -1,0 +1,18 @@
+// @flow
+
+// ----- Imports ----- //
+
+import { connect } from 'react-redux';
+
+import ContributionsThankYouPage from 'components/contributionsThankYou/contributionsThankYouPage';
+
+// ----- Map State/Props ----- //
+
+function mapStateToProps() {
+  return {
+    contributionType: 'ONE_OFF',
+    isDirectDebit: false,
+  };
+}
+
+export default connect(mapStateToProps)(ContributionsThankYouPage);
